+++ src/powercrystals/minefactoryreloaded/render/entity/EntityNeedleRenderer.java
@@ -22,14 +22,16 @@ public class EntityNeedleRenderer extends Render
         GL11.glRotatef(needle.prevRotationPitch + (needle.rotationPitch - needle.prevRotationPitch) * par9, 0.0F, 0.0F, 1.0F);
         Tessellator tessellator = Tessellator.instance;
         byte b0 = 0;
-        float f2 = 0.0F;
-        float f3 = 0.5F;
-        float f4 = (float)(0 + b0 * 10) / 32.0F;
-        float f5 = (float)(5 + b0 * 10) / 32.0F;
-        float f6 = 0.0F;
-        float f7 = 0.15625F;
-        float f8 = (float)(5 + b0 * 10) / 32.0F;
-        float f9 = (float)(10 + b0 * 10) / 32.0F;
+        
+        double f2 = 0.0D;
+        double f3 = 0.5D;
+        double f4 = (0 + b0 * 10) / 32.0D;
+        double f5 = (5 + b0 * 10) / 32.0D;
+        
+        double f6 = 0.0D;
+        double f7 = 0.15625D;
+        double f8 = (5 + b0 * 10) / 32.0D;
+        double f9 = (10 + b0 * 10) / 32.0D;
         float f10 = 0.01F;
         GL11.glEnable(GL12.GL_RESCALE_NORMAL);
 
@@ -38,17 +40,17 @@ public class EntityNeedleRenderer extends Render
         GL11.glTranslatef(-4.0F, 0.0F, 0.0F);
         GL11.glNormal3f(f10, 0.0F, 0.0F);
         tessellator.startDrawingQuads();
-        tessellator.addVertexWithUV(-7.0D, -2.0D, -2.0D, (double)f6, (double)f8);
-        tessellator.addVertexWithUV(-7.0D, -2.0D, 2.0D, (double)f7, (double)f8);
-        tessellator.addVertexWithUV(-7.0D, 2.0D, 2.0D, (double)f7, (double)f9);
-        tessellator.addVertexWithUV(-7.0D, 2.0D, -2.0D, (double)f6, (double)f9);
+        tessellator.addVertexWithUV(-7.0D, -2.0D, -2.0D, f6, f8);
+        tessellator.addVertexWithUV(-7.0D, -2.0D, 2.0D, f7, f8);
+        tessellator.addVertexWithUV(-7.0D, 2.0D, 2.0D, f7, f9);
+        tessellator.addVertexWithUV(-7.0D, 2.0D, -2.0D, f6, f9);
         tessellator.draw();
         GL11.glNormal3f(-f10, 0.0F, 0.0F);
         tessellator.startDrawingQuads();
-        tessellator.addVertexWithUV(-7.0D, 2.0D, -2.0D, (double)f6, (double)f8);
-        tessellator.addVertexWithUV(-7.0D, 2.0D, 2.0D, (double)f7, (double)f8);
-        tessellator.addVertexWithUV(-7.0D, -2.0D, 2.0D, (double)f7, (double)f9);
-        tessellator.addVertexWithUV(-7.0D, -2.0D, -2.0D, (double)f6, (double)f9);
+        tessellator.addVertexWithUV(-7.0D, 2.0D, -2.0D, f6, f8);
+        tessellator.addVertexWithUV(-7.0D, 2.0D, 2.0D, f7, f8);
+        tessellator.addVertexWithUV(-7.0D, -2.0D, 2.0D, f7, f9);
+        tessellator.addVertexWithUV(-7.0D, -2.0D, -2.0D, f6, f9);
         tessellator.draw();
 
         for (int i = 0; i < 4; ++i)
@@ -56,10 +58,10 @@ public class EntityNeedleRenderer extends Render
             GL11.glRotatef(90.0F, 1.0F, 0.0F, 0.0F);
             GL11.glNormal3f(0.0F, 0.0F, f10);
             tessellator.startDrawingQuads();
-            tessellator.addVertexWithUV(-8.0D, -2.0D, 0.0D, (double)f2, (double)f4);
-            tessellator.addVertexWithUV(8.0D, -2.0D, 0.0D, (double)f3, (double)f4);
-            tessellator.addVertexWithUV(8.0D, 2.0D, 0.0D, (double)f3, (double)f5);
-            tessellator.addVertexWithUV(-8.0D, 2.0D, 0.0D, (double)f2, (double)f5);
+            tessellator.addVertexWithUV(-8.0D, -2.0D, 0.0D, f2, f4);
+            tessellator.addVertexWithUV(8.0D, -2.0D, 0.0D, f3, f4);
+            tessellator.addVertexWithUV(8.0D, 2.0D, 0.0D, f3, f5);
+            tessellator.addVertexWithUV(-8.0D, 2.0D, 0.0D, f2, f5);
             tessellator.draw();
         }
